Some more va_list stuff; actually, my previous patch was incomplete (submitted over 18 months ago ...). Here comes the fix; the cvs version should work again with with patch, even on x86 and not only on x86_64 ...
See

http://gcc.gnu.org/bugzilla/show_bug.cgi?id=14557

The problem is that va_list is not only opaque, but ReallyOpaque
(maybe (tm)??). va_list may even be an array_type. Then in the context
of

foo(va_list foo)
{}

already is not of type "va_list" _BUT_ of type "&va_list[0]". So:

foo(va_list foo)
{
	bar(&foo);
}

passes _NOT_ the base address of va_list to "bar()" _BUT_ the address
of the piece of storage which holds the parameter "foo" which is _NOT_
what we want (think about it!! This is JAIOCABD
(JustAnotherInstanceOfCArrayBrainDamegedness) Oops.

I did all this to make Geomview work on AMD64 with Linux. Actually, I
DO NOT like the libc implementation of va_list, but we have to cope
with it. Also, the changes _should_ not break archs/compilers where
va_list is a structure and not of array-type.

Hope this will be portable (_should_ be, but FIXME if not).